nimble/ll: Rework how TX power is configured
TX power is now always configured after channel (and PHY) is set before
transmission is started. There is no "default" TX power and all users
are expected to set it explicitly before transmission. This means that
at the end of advertising TX power is not set back to global value.

Main rationale for this change is to properly handle FEM that may have
amplification factor depends on signal frequency (channel used).